Revert code block test
Using syntax highlighting as described at https://www.markdownguide.org/extended-syntax/#syntax-highlighting works - for a description of how to invoke it in GitBook or in the raw Markdown code, see https://docs.gitbook.com/editing-content/markdown#code-blocks and use xml as the language. Note that the example code block will render with a dark background if you do this, which may or may not be what you want, so will leave this as a plain code block without syntax highlighting for now.